# Improve StyledComponent performance