--- a/src/utils/generateAlphabeticName.js
+++ b/src/utils/generateAlphabeticName.js
@@ -1,12 +1,21 @@
 // @flow
 const chars = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ'.split('')
+const charsLength = chars.length
 
 /* Some high number, usually 9-digit base-10. Map it to base-😎 */
 const generateAlphabeticName = (code: number): string => {
-  const lastDigit = chars[code % chars.length]
-  return code > chars.length
-    ? `${generateAlphabeticName(Math.floor(code / chars.length))}${lastDigit}`
-    : lastDigit
+  let name = ''
+  let x
+
+  for (
+    x = code;
+    x > charsLength;
+    x = Math.floor(x / chars.length)
+  ) {
+    name = chars[x % charsLength] + name
+  }
+
+  return chars[x % charsLength] + name
 }
 
 export default generateAlphabeticName