#583

- Add second `BaseStyledComponent` class to not repeat methods
- Improve `propsForElement` and `this.attrs` creation
- Stop creating `this.attrs` if `attrs` doesn't exist
- Rewrite `generateAlphabeticName` to be iterative instead of recursive

### Before

```
simple component (construction)     0.019±0.003ms
simple component (first render)     0.132±0.027ms
simple component (repeated render)  0.014±0.001ms
prop changes                        0.309±0.054ms
prop shorthands (first render)      0.175±0.071ms
prop shorthands with prop changes   0.377±0.093ms
```


### After

```
simple component (construction)     0.013±0.002ms
simple component (first render)     0.121±0.026ms
simple component (repeated render)  0.015±0.004ms
prop changes                        0.311±0.101ms
prop shorthands (first render)      0.155±0.052ms
prop shorthands with prop changes   0.344±0.038ms
```

Edit:

### After with preprocessing

```
simple component (construction)     0.011±0.002ms
simple component (first render)     0.090±0.018ms
simple component (repeated render)  0.014±0.001ms
prop changes                        0.195±0.030ms
prop shorthands (first render)      0.095±0.030ms
prop shorthands with prop changes   0.208±0.051ms
```