@evan-scott-zocdoc 2017-03-31
Side comment, there is a potential performance benefit from changing the implementation of https://github.com/styled-components/styled-components/blob/d096d6126b606a5c413e353d2d6cedf2309a912d/src/models/AbstractStyledComponent.js#L6 to use `PureComponent` instead