@philpl 2017-03-31
@evan-scott-zocdoc That is unfortunately not quite the right way. First of all, it will change the behaviour for the user, and unexpected that is. And also, it doesn't improve the performance of the library itself. It's the responsibility of the user to decide when to turn to reference equality checks for perf optimisations.

The more clever approach is to not trigger state changes, when it's not necessary, but that's not really the scope of this PR. Btw: https://twitter.com/dan_abramov/status/820668074223353858