@philpl 2017-03-31
@mxstbr possibly! Next I want to see if there are more obvious things, maybe even related to behaviour as I've said above