@evan-scott-zocdoc 2017-04-01
@philpl how are you generating those benchmark stats? would be awesome to just have an npm script we could run when making changes to the library to quickly evaluate perf differences