@philpl 2017-04-01
@evan-scott-zocdoc With https://github.com/styled-components/benchmark

I've split up some tests into more granular ones though. Once we do #588, it's going to live in the main repo.